# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Health Tracker \u2014 API Gateway",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/telemetry": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "List Points",
        "operationId": "list_points_api_v1_telemetry_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "e.g. steps | stress | spo2 | heart_rate | energy_expenditure",
              "title": "Metric"
            },
            "description": "e.g. steps | stress | spo2 | heart_rate | energy_expenditure"
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO start (inclusive).",
              "title": "From"
            },
            "description": "ISO start (inclusive)."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO end (inclusive).",
              "title": "To"
            },
            "description": "ISO end (inclusive)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50000,
              "minimum": 1,
              "default": 1000,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TelemetryPoint"
                  },
                  "title": "Response List Points Api V1 Telemetry Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/daily": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Daily Rollup",
        "operationId": "daily_rollup_api_v1_telemetry_daily_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "metric",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Metric"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DailyRollup"
                  },
                  "title": "Response Daily Rollup Api V1 Telemetry Daily Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/telemetry/sleep": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Sleep Series",
        "description": "Per-night sleep summaries (total + stages) for the sleep trend.",
        "operationId": "sleep_series_api_v1_telemetry_sleep_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO start (inclusive).",
              "title": "From"
            },
            "description": "ISO start (inclusive)."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO end (inclusive).",
              "title": "To"
            },
            "description": "ISO end (inclusive)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SleepNight"
                  },
                  "title": "Response Sleep Series Api V1 Telemetry Sleep Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/body-composition": {
      "get": {
        "tags": [
          "telemetry"
        ],
        "summary": "Body Composition",
        "description": "Smart-scale series (weight, body fat %, skeletal muscle, BMR) for trends.",
        "operationId": "body_composition_api_v1_body_composition_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO start (inclusive).",
              "title": "From"
            },
            "description": "ISO start (inclusive)."
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO end (inclusive).",
              "title": "To"
            },
            "description": "ISO end (inclusive)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BodyCompositionPoint"
                  },
                  "title": "Response Body Composition Api V1 Body Composition Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendations",
        "description": "Today's recommendations. Generated + stored lazily on first request for a day.",
        "operationId": "get_recommendations_api_v1_recommendations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Day to fetch (YYYY-MM-DD, UTC).",
              "title": "Date"
            },
            "description": "Day to fetch (YYYY-MM-DD, UTC)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/run": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Run Recommendations",
        "description": "Force a regenerate + store (the daily-pass trigger).",
        "operationId": "run_recommendations_api_v1_recommendations_run_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "description": "Day to (re)generate.",
              "title": "Date"
            },
            "description": "Day to (re)generate."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/feedback": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Submit Feedback",
        "operationId": "submit_feedback_api_v1_recommendations_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BodyCompositionPoint": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "skeletal_muscle_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skeletal Muscle Kg"
          },
          "bmr_kcal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bmr Kcal"
          }
        },
        "type": "object",
        "required": [
          "ts"
        ],
        "title": "BodyCompositionPoint",
        "description": "One smart-scale reading (sparse periodic series)."
      },
...
      "DailyRollup": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "avg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg"
          },
          "min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min"
          },
          "max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max"
          },
          "sum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sum"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "day",
          "metric",
          "count"
        ],
        "title": "DailyRollup",
        "description": "One day's rollup for a metric. `sum` is meaningful for additive metrics\n(steps, energy_expenditure); `avg`/`min`/`max` for the rest."
      },
...
      "FeedbackIn": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "feedback": {
            "type": "string",
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "date",
          "feedback"
        ],
        "title": "FeedbackIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RecommendationItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "signals": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Signals",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "code",
          "category",
          "severity",
          "title",
          "detail"
        ],
        "title": "RecommendationItem"
      },
      "RecommendationsOut": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationItem"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "context": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "date",
          "recommendations"
        ],
        "title": "RecommendationsOut"
      },
      "SleepNight": {
        "properties": {
          "total_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Min"
          },
          "deep_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deep Min"
          },
          "rem_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rem Min"
          },
          "light_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Light Min"
          },
          "awake_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awake Min"
          },
          "efficiency": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Efficiency"
          },
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          }
        },
        "type": "object",
        "required": [
          "day"
        ],
        "title": "SleepNight",
        "description": "A night's sleep summary tagged with the calendar day it ended on (for trends)."
      },
...
      "TelemetryPoint": {
        "properties": {
          "ts": {
            "type": "string",
            "format": "date-time",
            "title": "Ts"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "ts",
          "metric",
          "value",
          "source"
        ],
        "title": "TelemetryPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Shared gateway token (no 'Bearer ' prefix).",
        "scheme": "bearer"
      }
    }
  }
}